requirements.txt
annotated-types==0.7.0
anyio==4.10.0
click==8.2.1
fastapi==0.116.1
h11==0.16.0
idna==3.10
pydantic==2.11.7
pydantic_core==2.33.2
sniffio==1.3.1
starlette==0.47.3
typing-inspection==0.4.1
typing_extensions==4.15.0
uvicorn==0.35.0
fastapi>=0.111.0
uvicorn[standard]>=0.30.0
pydantic>=2.7.0
pymongo>=4.7.0
langgraph>=0.2.21
langchain-core>=0.2.39
langchain-openai>=0.1.21
streamlit
langgraph-checkpoint-mongodb

fastapi
uvicorn[standard]
langgraph
langchain-openai
openai
pydantic
pymongo
torch
clip
pillow
chromadb
scikit-learn
python-multipart

